chore: remove fully qualified reference
