+++ use-cases/bed-allocation/src/main/java/org/acme/bedallocation/demo/BedAllocationDemoDataGenerator.java
@@ -1,32 +1,33 @@
-package org.acme.conferencescheduling.demo;
+package org.acme.bedallocation.demo;
 
 import java.util.List;
 
 import jakarta.enterprise.context.ApplicationScoped;
 
 import ai.timefold.solver.service.definition.api.data.DemoData;
+import ai.timefold.solver.service.definition.api.data.DemoDataGenerator;
 import ai.timefold.solver.service.definition.api.data.DemoMetaData;
 import ai.timefold.solver.service.definition.api.domain.Configuration;
 import ai.timefold.solver.service.definition.api.domain.ModelConfig;
 import ai.timefold.solver.service.definition.api.domain.ModelRequest;
 import ai.timefold.solver.service.definition.api.domain.RunConfiguration;
 
-import org.acme.conferencescheduling.dto.input.ConferenceScheduleConfigOverrides;
-import org.acme.conferencescheduling.dto.input.ConferenceScheduleInput;
+import org.acme.bedallocation.dto.input.BedPlanConfigOverrides;
+import org.acme.bedallocation.dto.input.BedPlanInput;
 
 @ApplicationScoped
-public class DemoDataGenerator implements ai.timefold.solver.service.definition.api.data.DemoDataGenerator {
+public class BedAllocationDemoDataGenerator implements DemoDataGenerator {
 
     private static final String BASIC_DEMO_DATA_ID = "BASIC";
 
     private static final DemoMetaData BASIC_META_DATA = new DemoMetaData(BASIC_DEMO_DATA_ID,
-            "Demonstrates a conference scheduling problem with talk, speaker, room, and audience tags.",
-            "Schedules 20 talks by 12 speakers into 6 timeslots across 7 rooms. Talks and speakers carry theme, "
-                    + "sector, audience, and content tags that drive requirements, preferences, diversity, and "
-                    + "conflict rules. Once solved, talks are placed so speaker and room conflicts are avoided "
-                    + "and tag-based preferences (required rooms, preferred timeslots, audience-level flow) are "
-                    + "honored.",
-            List.of("talk tags", "speaker tags", "room tags", "theme tracks", "audience diversity"),
+            "Demonstrates a hospital bed allocation problem with room capacity, equipment, and gender rules.",
+            "Assigns patient stays to beds in a single department of 10 rooms over a 28-day horizon. Stays carry "
+                    + "a required specialty and preferred equipment (telemetry, television, oxygen, nitrogen), and "
+                    + "rooms enforce gender-mix rules across nights. Once solved, every patient is assigned to a "
+                    + "bed that respects gender, age, specialty, and equipment preferences while keeping stays in "
+                    + "the same bed night to night.",
+            List.of("room capacity", "equipment", "gender rules", "specialty", "age"),
             List.of());
 
     @Override
@@ -39,9 +40,8 @@ public class DemoDataGenerator implements ai.timefold.solver.service.definition.
         if (!BASIC_DEMO_DATA_ID.equals(id)) {
             throw new IllegalArgumentException("Unknown demo data id (%s).".formatted(id));
         }
-        ConferenceScheduleInput problem = DemoDataBuilder.builder().build();
-
-        Configuration<ConferenceScheduleConfigOverrides> configuration = new Configuration<>(
+        BedPlanInput problem = DemoDataBuilder.builder().build();
+        Configuration<BedPlanConfigOverrides> configuration = new Configuration<>(
                 new RunConfiguration(BASIC_DEMO_DATA_ID), ModelConfig.empty());
         return new DemoData(BASIC_META_DATA, new ModelRequest<>(configuration, problem));
     }
